refactor(api): split src/lib/api.ts into domain-scoped modules (CHAOS-1219)
Decomposes the 517-LOC `src/lib/api.ts` facade into 6 domain modules under
`src/lib/api/` plus a private shared-helpers module. Deletes the barrel
entirely — all 27 consumer files migrated to deep imports.

## New modules

- `src/lib/api/_shared.ts` — private helpers (`normalizeFilters`, `postJson`,
  `getAuth`). Underscore-prefixed to signal module-private intent.
- `src/lib/api/home.ts` — `getHomeData`, `getExplainData`, `getOpportunities`
- `src/lib/api/investment.ts` — `getInvestment`, `explainInvestmentMix`,
  `getSankey`, `getInvestmentFlow`, `getInvestmentRepoTeamFlow`, `getWorkUnits`,
  `getWorkUnitExplanation`, `getDrilldown`
- `src/lib/api/people.ts` — `searchPeople`, `getPersonSummary`,
  `getPersonMetric`, `getPersonDrilldown`
- `src/lib/api/visuals.ts` — `getHeatmap`, `getFlame`, `getAggregatedFlame`,
  `getQuadrant`
- `src/lib/api/system.ts` — `checkApiHealth`, `getApiMeta`
- `src/lib/api/capacity.ts` — `getCapacityForecast`

## No new barrel

Deliberately NO `src/lib/api/index.ts`. Per the lesson learned in CHAOS-1236
(#433), barrels transitively pull server-only/heavy modules into client
bundles. Consumers import the specific domain they need.

## Migration

All 27 importers (22 pages + 5 components + 1 test file) migrated to their
appropriate deep paths. Multi-domain consumers (e.g. `work/page.tsx`,
`code/page.tsx`, `explore/page.tsx`) now have 2-3 import lines instead of one
grab-bag. The domain separation makes it easier to reason about what each page
actually touches.

## Zero behavior change

This is a pure code move. Every function kept its exact body, signature,
inline comments, and error semantics. The only non-move changes are import
paths. Verified by:
- `pnpm typecheck` — 0 errors
- `pnpm test:unit` — 867/867 passing (same count as pre-split)
- `pnpm lint` — clean

Closes CHAOS-1219. Refs CHAOS-1211 (Data Layer epic).
