+++ src/app/(app)/dashboard/page.tsx
@@ -7,7 +7,8 @@ import { ServiceUnavailable } from "@/components/ServiceUnavailable";
 import { FilterBar } from "@/components/filters/FilterBar";
 import { PrimaryNav } from "@/components/navigation/PrimaryNav";
 import { RoleSelectorWithSuspense, RoleFraming } from "@/components/RoleSelectorWrapper";
-import { checkApiHealth, getApiMeta, getHomeData } from "@/lib/api";
+import { checkApiHealth, getApiMeta } from "@/lib/api/system";
+import { getHomeData } from "@/lib/api/home";
 import { decodeFilter, filterFromQueryParams } from "@/lib/filters/encode";
 import { buildExploreUrl, withFilterParam } from "@/lib/filters/url";
 import { ClientTimestamp } from "@/components/ClientTimestamp";
