+++ src/app/(app)/investment/page.tsx
@@ -5,7 +5,8 @@ import { FilterBar } from "@/components/filters/FilterBar";
 import { InvestmentChart } from "@/components/investment/InvestmentChart";
 import { PrimaryNav } from "@/components/navigation/PrimaryNav";
 import { ServiceUnavailable } from "@/components/ServiceUnavailable";
-import { checkApiHealth, getInvestment } from "@/lib/api";
+import { checkApiHealth } from "@/lib/api/system";
+import { getInvestment } from "@/lib/api/investment";
 import { getCurrentOrg, getOrgEntitlements } from "@/lib/admin/server";
 import { fetchOrNull } from "@/lib/fetchOrNull";
 import { decodeFilter, filterFromQueryParams } from "@/lib/filters/encode";
