+++ src/app/(app)/metrics/page.tsx
@@ -6,7 +6,9 @@ import { FilterBar } from "@/components/filters/FilterBar";
 import { MetricCard } from "@/components/metrics/MetricCard";
 import { PrimaryNav } from "@/components/navigation/PrimaryNav";
 import { ServiceUnavailable } from "@/components/ServiceUnavailable";
-import { checkApiHealth, getExplainData, getHomeData, getQuadrant } from "@/lib/api";
+import { checkApiHealth } from "@/lib/api/system";
+import { getExplainData, getHomeData } from "@/lib/api/home";
+import { getQuadrant } from "@/lib/api/visuals";
 import { decodeFilter, filterFromQueryParams } from "@/lib/filters/encode";
 import { fetchOrNull } from "@/lib/fetchOrNull";
 import { buildExploreUrl, withFilterParam } from "@/lib/filters/url";
