+++ src/app/(app)/opportunities/page.tsx
@@ -3,7 +3,8 @@ import Link from "next/link";
 import { FilterBar } from "@/components/filters/FilterBar";
 import { PrimaryNav } from "@/components/navigation/PrimaryNav";
 import { ServiceUnavailable } from "@/components/ServiceUnavailable";
-import { checkApiHealth, getOpportunities } from "@/lib/api";
+import { checkApiHealth } from "@/lib/api/system";
+import { getOpportunities } from "@/lib/api/home";
 import { decodeFilter, filterFromQueryParams } from "@/lib/filters/encode";
 import { fetchOrNull } from "@/lib/fetchOrNull";
 import { buildExploreUrl, withFilterParam } from "@/lib/filters/url";
