+++ src/app/(app)/people/[person_id]/metrics/[metric]/page.tsx
@@ -5,7 +5,9 @@ import { HeatmapPanel } from "@/components/charts/HeatmapPanel";
 import { TimeseriesChart } from "@/components/charts/TimeseriesChart";
 import { PersonRangeBar } from "@/components/people/PersonRangeBar";
 import { PrimaryNav } from "@/components/navigation/PrimaryNav";
-import { checkApiHealth, getHeatmap, getPersonDrilldown, getPersonMetric, getPersonSummary } from "@/lib/api";
+import { checkApiHealth } from "@/lib/api/system";
+import { getHeatmap } from "@/lib/api/visuals";
+import { getPersonDrilldown, getPersonMetric, getPersonSummary } from "@/lib/api/people";
 import { defaultMetricFilter } from "@/lib/filters/defaults";
 import { fetchOrNull } from "@/lib/fetchOrNull";
 import { decodeFilter } from "@/lib/filters/encode";
