+++ src/app/(app)/people/[person_id]/page.tsx
@@ -6,7 +6,9 @@ import { QuadrantPanel } from "@/components/charts/QuadrantPanel";
 import { PersonRangeBar } from "@/components/people/PersonRangeBar";
 import { MetricCard } from "@/components/metrics/MetricCard";
 import { PrimaryNav } from "@/components/navigation/PrimaryNav";
-import { checkApiHealth, getPersonSummary, getQuadrant } from "@/lib/api";
+import { checkApiHealth } from "@/lib/api/system";
+import { getPersonSummary } from "@/lib/api/people";
+import { getQuadrant } from "@/lib/api/visuals";
 import { defaultMetricFilter } from "@/lib/filters/defaults";
 import { fetchOrNull } from "@/lib/fetchOrNull";
 import { decodeFilter } from "@/lib/filters/encode";
