+++ src/app/(app)/prs/[pr_id]/page.tsx
@@ -3,7 +3,8 @@ import Link from "next/link";
 import { FlameDiagram } from "@/components/charts/FlameDiagram";
 import { PrimaryNav } from "@/components/navigation/PrimaryNav";
 import { ServiceUnavailable } from "@/components/ServiceUnavailable";
-import { checkApiHealth, getFlame } from "@/lib/api";
+import { checkApiHealth } from "@/lib/api/system";
+import { getFlame } from "@/lib/api/visuals";
 import { defaultMetricFilter } from "@/lib/filters/defaults";
 import { ClientTimestamp } from "@/components/ClientTimestamp";
 import { fetchOrNull } from "@/lib/fetchOrNull";
